chore: move utils directory into lib and restored the functionality of the emptyButton
