+++ rs-soap-shop/src/components/buttons/emptyButton/index.tsx
@@ -1,14 +1,12 @@
 import { ButtonProps } from '../../../lib/interfaces'
-import {Link} from 'react-router-dom';
 
-export default function EmptyButton({ children, to, onClick }: ButtonProps): JSX.Element {
+export default function EmptyButton({ children, onClick }: ButtonProps): JSX.Element {
   return (
-    <Link
+    <button
       onClick={onClick}
-      to={to}
-      className='transition inline-flex whitespace-nowrap items-center text-accentColor dark:text-secondaryColor border-2 border-accentColor dark:border-secondaryColor hover:text-secondaryColor dark:hover:text-grayLColor font-bold bg-none hover:bg-accentColor dark:hover:bg-secondaryColor rounded-normal h-[74px] px-12 active:scale-95'
+      className='transition text-accentColor dark:text-secondaryColor border-2 border-accentColor dark:border-secondaryColor hover:text-secondaryColor dark:hover:text-grayLColor font-bold bg-none hover:bg-accentColor dark:hover:bg-secondaryColor rounded-normal h-[74px] px-12 active:scale-95'
     >
       {children}
-    </Link>
+    </button>
   )
 }
