+++ rs-soap-shop/src/lib/utils/findInputError.ts
