+++ rs-soap-shop/src/lib/utils/isFormInvalid.ts
@@ -1,4 +1,4 @@
-import { InputErrorObject } from '../lib/interfaces'
+import { InputErrorObject } from '../interfaces'
 
 export const isFormInvalid = (err: InputErrorObject) => {
   if (Object.keys(err).length > 0) return true
